feat(editimagemodal): edit image modal to 2 pages
add AssetsList, Pagination to EditImageModal

Add an AssetsList component to provide a selectable list of assets in the EditImageModal. Add AssetsList and existing Pagination component to EditImageModal.

add assets dropzone to image modal

add search and result count (#196)

Wraps the result count so that it is not aware of the Images filter that is
always applied by default.

NoAssets and NoResults views, clear search button