perf(assets): use paragon pagination & search
Paragon's `Pagination` and `SearchField` components are now used for
the Assets page and the parallel Add Image Modal. This involves some
visual changes, and a difference in the number of pages ellided in the
navigation bar.

`perf` in the title here is used to force a new patch version, but
`refactor` would likely be more descriptive.

fixes EDUCATOR-3421